Rewrite the entire fucking thing in Python...
because NodeJS was never built to do the task and I really didn't want to write it in Python then.
I still don't like Python but holy hell is it so much better at the task than Node ever could be.
The entire rewrite was a breeze compared to the living hell that Node was.